+++ src/query_ids.js
@@ -12,6 +12,7 @@ const QUERY_IDS = {
     postShortCodeMedia: '2efa04f61586458cef44441f474eee7c',
     profileHighlights: '45246d3fe16ccc6577e0bd297a5db1ab',
     profilePublicStories: 'd4d88dc1500312af6f937f7b804c68c3',
+    taggedPosts: 'be13233562af2d229b008d2976b998b5',
     profileStories: 'c9c56db64beb4c9dea2d17740d0259d9',
 };